apps: move dispatch code to another mixin
