+++ apps/contrib/mixins.py
@@ -1,5 +1,7 @@
 from django import forms
+from django.utils.functional import cached_property
 from django.utils.translation import ugettext_lazy as _
+from django.views import generic
 
 RIGHT_OF_USE_LABEL = _('I hereby confirm that the copyrights for this '
                        'photo are with me or that I have received '
@@ -58,3 +60,36 @@ class ImageRightOfUseMixin(forms.ModelForm):
                            _("You want to upload an image. "
                              "Please check that you have the "
                              "right of use for the image."))
+
+
+class ProjectModuleDispatchMixin(generic.DetailView):
+
+    @cached_property
+    def project(self):
+        return self.get_object()
+
+    @cached_property
+    def module(self):
+        if self.modules.count() == 1 and not self.events:
+            return self.modules.first()
+        elif len(self.get_current_modules()) == 1:
+            return self.get_current_modules()[0]
+
+    def dispatch(self, request, *args, **kwargs):
+        kwargs['project'] = self.project
+        kwargs['module'] = self.module
+
+        if self.modules.count() == 1 and not self.events:
+            return self._view_by_phase()(request, *args, **kwargs)
+        elif len(self.get_current_modules()) == 1:
+            return self._view_by_phase()(request, *args, **kwargs)
+        else:
+            return super().dispatch(request)
+
+    def _view_by_phase(self):
+        if self.module.last_active_phase:
+            return self.module.last_active_phase.view.as_view()
+        elif self.module.future_phases:
+            return self.module.future_phases.first().view.as_view()
+        else:
+            return super().dispatch
